+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.2.0]
+
+### New Features
+- Now supporting Netzsch TGA files
+
+### Fixes/Minor Changes
+- Mettler Toledo files: Added default units and fixed DTGA calculation
+- PE files: Fixed Date and Time parsing for ANSI encoded PE files
+- the name of cracking and burnoff stages for pyro experiments can now be simply passed to the parsing function
+
 ## [0.1.1] - 2025-06-19
 
 - Initial PyPI release